More bug fixes
- Changed the CPU Usage type to a float and added support for floats on the MQTT side as the number parsed was a floating point value

+++ gwn/api/GwnClient.py
@@ -152,7 +152,7 @@ class GwnClient:
                         temperature_c=int(config_info_client["temperature"].replace("℃", "").replace("°C", "").strip()),
                         usedMemory_bytes=self._size_to_bytes(config_info_client["usedMemory"]),
                         channelload_2g4_percent=int(config_info_client["channelload_2g4"]) if config_info_client["channelload_2g4"] is not None else 0,
-                        cpuUsage_percent=int(config_info_client["cpuUsage"].replace("%","").strip()) if config_info_client["cpuUsage"] is not None else 0,
+                        cpuUsage_percent=float(config_info_client["cpuUsage"].replace("%","").strip()) if config_info_client["cpuUsage"] is not None else 0,
                         channelload_6g_percent=int(config_info_client["channelload_6g"]) if config_info_client["channelload_6g"] is not None else 0,
                         channelload_5g_percent=int(config_info_client["channelload_5g"]) if config_info_client["channelload_5g"] is not None else 0,
